Use `ILspEditorFeatureDetector` interface rather than an abstract class
